Move extra tests to separate file and import them.
